refactor: exception handler 방식 변경
