--- src/main/java/org/poolc/api/auth/exception/AuthExceptionHandlers.java
@@ -1,25 +1,0 @@
-package org.poolc.api.auth.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import java.util.Collections;
-import java.util.Map;
-
-@ControllerAdvice("org.poolc.api.auth")
-public class AuthExceptionHandlers {
-
-    @ExceptionHandler({UnactivatedException.class, UnauthorizedException.class})
-    public ResponseEntity<Map<String, String>> unactivatedMemberHandler(Exception e) {
-        return ResponseEntity.status(HttpStatus.FORBIDDEN)
-                .body(Collections.singletonMap("data", e.getMessage()));
-    }
-
-    @ExceptionHandler(UnauthenticatedException.class)
-    public ResponseEntity<Map<String, String>> unauthenticatedMemberHandler(Exception e) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                .body(Collections.singletonMap("data", e.getMessage()));
-    }
-}
