+++ src/main/java/org/poolc/api/member/exception/MemberExceptionHandlers.java
@@ -1,21 +1,18 @@
 package org.poolc.api.member.exception;
 
-import javassist.bytecode.DuplicateMemberException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
 
-@ControllerAdvice("poolc.poolc.member.exception")
+import java.util.Collections;
+import java.util.Map;
+
+@ControllerAdvice("org.poolc.api.member")
 public class MemberExceptionHandlers {
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<String> incorrectInputHandler(Exception e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ResponseStatus(HttpStatus.CONFLICT)
     @ExceptionHandler(DuplicateMemberException.class)
-    public void memberAlreadyExistException() {
+    public ResponseEntity<Map<String, String>> memberAlreadyExistException(Exception e) {
+        return ResponseEntity.status(HttpStatus.CONFLICT)
+                .body(Collections.singletonMap("message", e.getMessage()));
     }
 }
