+++ src/main/java/org/poolc/api/post/controller/PostController.java
@@ -1,7 +1,6 @@
 package org.poolc.api.post.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.poolc.api.auth.exception.UnauthenticatedException;
 import org.poolc.api.board.domain.Board;
 import org.poolc.api.board.service.BoardService;
 import org.poolc.api.member.domain.Member;
@@ -14,7 +13,6 @@ import org.poolc.api.post.dto.UpdatePostRequest;
 import org.poolc.api.post.service.PostService;
 import org.poolc.api.post.vo.PostCreateValues;
 import org.poolc.api.post.vo.PostUpdateValues;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -23,7 +21,6 @@ import org.springframework.web.bind.annotation.*;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 
 @RestController
@@ -109,13 +106,6 @@ public class PostController {
         return ResponseEntity.ok().build();
     }
 
-    private void checkWriter(Member user, Post updatePost) {
-        if (!updatePost.getMember().getUUID().equals(user.getUUID())) {
-            throw new org.poolc.api.auth.exception.UnauthorizedException("접근할 수 없습니다.");
-        }
-    }
-
-
     @DeleteMapping(value = "/{postId}")
     public ResponseEntity<Void> deletePost(@AuthenticationPrincipal Member user, @PathVariable Long postId) {
         Post updatePost = postService.getPost(postId);
@@ -126,14 +116,10 @@ public class PostController {
         return ResponseEntity.ok().build();
     }
 
-    @ExceptionHandler({UnauthenticatedException.class, org.poolc.api.auth.exception.UnauthorizedException.class})
-    public ResponseEntity<String> unauthenticatedHandler(Exception e) {
-        return ResponseEntity.status(HttpStatus.FORBIDDEN).body(e.getMessage());
-    }
-
-    @ExceptionHandler({NoSuchElementException.class, IllegalArgumentException.class})
-    public ResponseEntity<String> noSuchElementHandler(Exception e) {
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
+    private void checkWriter(Member user, Post updatePost) {
+        if (!updatePost.getMember().getUUID().equals(user.getUUID())) {
+            throw new org.poolc.api.auth.exception.UnauthorizedException("접근할 수 없습니다.");
+        }
     }
 
     private void checkWritePermissions(Member user, Board board) {
